Rename SkillModificationParser to AdditionalSkillStatParser
